rename gd_resource_ssab to gd_ssab_resource
