+++ gd_spritestudio/gd_ssab_resource.h
